\name{contain_string}
\alias{contain_string}
\title{Check the transaction similarity of a user}
\usage{
contain_string(col1, col2)
}
\arguments{
  \item{col1}{character column contain string to match.}

  \item{col2}{character column contain string to match.}
}
\value{
A function consums a string and returns Boolean vector
indicating if there is a match in the input columns or not.
}
\description{
Check the transaction similarity of a user
}

