\name{eud.dist_center}
\alias{eud.dist_center}
\title{Calculate euclidean distance to the centeroid}
\usage{
eud.dist_center(x1)
}
\arguments{
  \item{x1}{data matrix, each row is one data record each
  column is one dimision.}
}
\value{
average euclidean distance with respect to the center point
of the data points.
}
\description{
Calculate euclidean distance to the centeroid
}

