\name{get_name_tag}
\alias{get_name_tag}
\title{Check the transaction similarity of a user}
\usage{
get_name_tag(col1, col2)
}
\arguments{
  \item{col1}{character column contain string to match.}

  \item{col2}{character column contain string to match.}
}
\value{
A function consums a list of strings and returns a
collection of strings in the list that matches the content
in either input columns.
}
\description{
Check the transaction similarity of a user
}

