\name{is_end_of_month}
\alias{is_end_of_month}
\title{Check if the date is the Last doy of the month}
\usage{
is_end_of_month(date)
}
\arguments{
  \item{date}{the date in date type.}
}
\value{
Boolean vectors indicating if the input is the last day of
the month or not
}
\description{
Check if the date is the Last doy of the month
}

