\name{is_start_of_month}
\alias{is_start_of_month}
\title{Check if the date is the First doy of the month}
\usage{
is_start_of_month(date)
}
\arguments{
  \item{date}{the date in date type.}
}
\value{
Boolean vectors indicating if the input is the first day of
the month or not
}
\description{
Check if the date is the First doy of the month
}

